requirements.txt
dash==2.17.1
dash-bootstrap-components==1.6.0

pandas==2.2.2
openpyxl==3.1.2

reportlab==4.2.0
num2words==0.5.13

sendgrid==6.11.0

openai==1.30.1

python-dotenv==1.0.1
requests==2.31.0